Validasi type file

+++ cms/photo/photo_domain/photo.go
@@ -8,7 +8,7 @@ import (
 type Photo struct {
 	ID         uint64         `json:"id" gorm:"primarykey"`
 	Images     string         `json:"images"`
-	CategoryID uint           `json:"category_id"`
+	CategoryID uint64         `json:"category_id"`
 	CreatedAt  time.Time      `json:"created_at"`
 	UpdatedAt  time.Time      `json:"updated_at"`
 	DeletedAt  gorm.DeletedAt `json:"deleted_at" gorm:"index"`
